chore(web): refactor storypanel (#685)
