+++ web/src/beta/lib/core/StoryPanel/Block/builtin/common/usePropertyValueUpdate.ts
