+++ web/src/beta/lib/core/StoryPanel/PageIndicator/index.stories.tsx
