+++ web/src/beta/lib/core/StoryPanel/SelectableArea/index.tsx
