+++ web/src/services/api/storytellingApi/blocks.ts
@@ -36,7 +36,7 @@ export type InstallableStoryBlock = {
   extensionId: string;
   icon?: string;
   singleOnly?: boolean;
-  type?: PluginExtensionType;
+  type?: "StoryBlock";
 };
 
 export type InstalledStoryBlock = {
@@ -176,7 +176,7 @@ const getInstallableStoryBlocks = (rawScene?: GetSceneQuery) => {
             extensionId: e.extensionId,
             icon: e.icon,
             singleOnly: !!e.singleOnly,
-            type: e.type,
+            type: "StoryBlock",
           };
         })
         .filter((sb): sb is InstallableStoryBlock => !!sb);
